+++ CHANGELOG.md
@@ -1,9 +1,15 @@
 # Changelog
 
+## 0.7.4
+
++ Delay release of `@glue()`
++ Add docs and some tests to `.iding`
++ Fix bug in `SiqGen()` that may prevent generation in short amounts of time
+
 ## 0.7.3
 
 + Fixed some broken imports in `.sqlalchemy`
-+ Stage `@glue()` for release in 0.8.0
++ Stage `@glue()` for release in 0.9.0
 + Add docs to `.sqlalchemy`
 
 ## 0.7.2